+++ src/main/java/Hampouch/server/domain/expense/dto/ExpenseDayListResponse.java
@@ -11,7 +11,7 @@ import java.util.List;
 /** 하루치 지출 목록 응답(GET /expenses/day). totalAmount를 함께 내려 일간 화면이 summary API를 또 호출하지 않게 한다. */
 public record ExpenseDayListResponse(
         LocalDate date,
-        int totalAmount,
+        long totalAmount,
         boolean hasRecord,
         List<ExpenseSummary> expenses
 ) {
@@ -41,7 +41,7 @@ public record ExpenseDayListResponse(
 
     /** totalAmount는 이미 조회한 리스트를 stream 합산 — 하루 단위라 비용 무시 가능(월/주 집계엔 재사용 금지). */
     public static ExpenseDayListResponse from(LocalDate date, List<Expense> expenses, boolean hasRecord) {
-        int totalAmount = expenses.stream().mapToInt(Expense::getPrice).sum();
+        long totalAmount = expenses.stream().mapToLong(Expense::getPrice).sum();
         List<ExpenseSummary> summaries = expenses.stream().map(ExpenseSummary::from).toList();
         return new ExpenseDayListResponse(date, totalAmount, hasRecord, summaries);
     }
