+++ src/main/java/Hampouch/server/domain/expense/service/PeriodSpending.java
@@ -7,7 +7,7 @@ import java.util.List;
  * totalAmount는 emotionBreakdown의 ratio 분모와 같은 값을 노출해 둔 것 — 응답과 문구가 다른 집계를 보지 않게 한다.
  */
 public record PeriodSpending(
-        int totalAmount,
+        long totalAmount,
         List<EmotionSpending> emotionBreakdown
 ) {
 }
